books-app/requirements.txt
alembic==1.14.0
annotated-types==0.7.0
anyio==4.7.0
click==8.1.8
exceptiongroup==1.2.2
fastapi==0.115.6
FastAPI-SQLAlchemy==0.2.1
graphql-core==3.2.5
greenlet==3.1.1
h11==0.14.0
idna==3.10
Mako==1.3.8
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
pydantic==2.10.4
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-multipart==0.0.20
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
strawberry-graphql==0.256.1
typing_extensions==4.12.2
uvicorn==0.34.0
